+++ CHANGELOG.md
@@ -15,6 +15,13 @@ Semantic versioning in our case means:
   There are no major releases right now: we are still at `0.x.y` version.
   But, in the future we might change the configuration names / logic,
   change the client facing API, change code conventions signigicantly, etc.
+  
+  
+## 0.15.3 WIP
+
+### Bugfixes
+
+- Fixes `UselessReturningElseViolation` to not report `else` with `break` #1958
 
 
 ## 0.15.2